third_party/WebKit/LayoutTests/inspector-protocol/css/css-get-platform-fonts.html

<html>
<head>
<script type="text/javascript" src="resources/inspector-protocol-test.js"></script>
<script type="text/javascript" src="resources/css-protocol-test.js"></script>
<script type="text/javascript" src="resources/dom-protocol-test.js"></script>
<script type="text/javascript">

function test()
{
    var documentNodeId;

    InspectorTest.requestDocumentNodeId(onDocumentNodeId);

    function onDocumentNodeId(nodeId)
    {
        documentNodeId = nodeId;
        InspectorTest.runTestSuite([
            function testFirstLetterPseudoClass(next)
            {
                platformFontsForElementWithSelector("#fancy", next);
            },

            function testSelectElementPlatformFonts(next)
            {
                platformFontsForElementWithSelector("select", next);
            }
        ]);
    }


    async function platformFontsForElementWithSelector(selector, callback)
    {
        var nodeId = await InspectorTest.requestNodeId(documentNodeId, selector);
        await InspectorTest.sendCommandOrDie("CSS.enable", {  });
        var response = await InspectorTest.sendCommandOrDie("CSS.getPlatformFontsForNode", { nodeId: nodeId });
        dumpComputedFonts(response);
        callback();
    }

    function dumpComputedFonts(response)
    {
        var fonts = response.fonts;
        fonts.sort(function(a, b) {
            return b.glyphCount - a.glyphCount;
        });
        for (var i = 0; i < fonts.length; ++i)
            fonts[i].familyName = "<Some-family-name-" + i + ">";
        for (var i = 0; i < fonts.length; ++i) {
            var lines = [
                "Font #" + i,
                "    name: " + fonts[i].familyName,
                "    glyphCount: " + fonts[i].glyphCount,
                "    isCustomFont: " + fonts[i].isCustomFont
            ];
            InspectorTest.log(lines.join("\n"));
        }
    }
};

window.addEventListener("DOMContentLoaded", function () {
    runTest();
}, false);

</script>
<style>
@font-face {
    font-family: "ahem";
    src: url('./resources/Ahem.ttf');
}

    #fancy {
        font-family: "ahem";
        background-color: gray;
    }
    #fancy:first-letter {
        font-family: 'Times New Roman';
        font-size: 400%;
        background-color: blue;
    }

    #fancy:first-line {
        font-family: 'Courier New';
        background-color: yellow;
    }
</style>
</head>
<body>

<div id="fancy">
7chars.<br>
Some line with 29 characters.
</div>
<select>
    <option>Short</option>
    <option selected>Option with a lot of chars.</option>
</select>

</body>
</html>
